requirements.txt
asgiref==3.11.0
Django==5.2.9
django-cors-headers==4.9.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
PyJWT==2.10.1
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.4
uritemplate==4.2.0
pytest
pytest-django
black
isort
channels==4.0.0
channels-redis==4.2.0